refactor: move parse function to file directory
