+++ src/Element/AmbientLight/ambient_light.c
@@ -6,7 +6,7 @@
 /*   By: yonshin <yonshin@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/09 12:04:44 by yonshin           #+#    #+#             */
-/*   Updated: 2023/07/03 22:39:05 by minjungk         ###   ########.fr       */
+/*   Updated: 2023/07/03 22:43:40 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,6 @@ t_func	ambient_light(enum e_element_func func)
 	static const t_func	functions[MAX_ELEMENT_TYPE] = {
 	[Init] = init_ambient_light,
 	[Hit] = hit_ambient_light,
-	[Parse] = parse_ambient_light,
 	[Debug] = debug_ambient_light};
 
 	return (functions[func]);
