+++ src/Element/Camera/camera.h
@@ -6,7 +6,7 @@
 /*   By: yonshin <yonshin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/31 05:36:34 by minjungk          #+#    #+#             */
-/*   Updated: 2023/06/10 01:51:03 by minjungk         ###   ########.fr       */
+/*   Updated: 2023/07/03 22:43:01 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,6 @@ typedef struct s_camera
 
 extern t_func	camera(enum e_element_func func);
 extern int		debug_camera(void *param);
-extern int		parse_camera(void *param, char **argv);
 extern int		hit_camera(void *this, t_ray *r, t_range ra, t_hit *h);
 extern int		init_camera(t_camera *self);
