+++ src/Element/Cylinder/cylinder.c
@@ -6,7 +6,7 @@
 /*   By: yonshin <yonshin@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/09 12:04:44 by yonshin           #+#    #+#             */
-/*   Updated: 2023/07/03 22:39:38 by minjungk         ###   ########.fr       */
+/*   Updated: 2023/07/03 22:43:52 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,6 @@ t_func	cylinder(enum e_element_func func)
 	static const t_func	functions[MAX_ELEMENT_TYPE] = {
 	[Init] = init_cylinder,
 	[Hit] = hit_cylinder,
-	[Parse] = parse_cylinder,
 	[Debug] = debug_cylinder};
 
 	return (functions[func]);
