+++ src/Element/Plane/plane.h
@@ -6,7 +6,7 @@
 /*   By: yonshin <yonshin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/31 05:38:22 by minjungk          #+#    #+#             */
-/*   Updated: 2023/06/10 01:51:34 by minjungk         ###   ########.fr       */
+/*   Updated: 2023/07/03 22:43:11 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,6 @@ typedef struct s_plane
 
 extern t_func	plane(enum e_element_func func);
 extern int		debug_plane(void *param);
-extern int		parse_plane(void *param, char **argv);
 extern int		hit_plane(void *this, t_ray *r, t_range ra, t_hit *h);
 extern int		init_plane(t_plane *self);
