+++ src/Element/Sphere/sphere.c
@@ -6,7 +6,7 @@
 /*   By: yonshin <yonshin@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/09 12:04:44 by yonshin           #+#    #+#             */
-/*   Updated: 2023/07/03 22:40:46 by minjungk         ###   ########.fr       */
+/*   Updated: 2023/07/03 22:44:13 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,6 @@ t_func	sphere(enum e_element_func func)
 	static const t_func	functions[MAX_ELEMENT_TYPE] = {
 	[Init] = init_sphere,
 	[Hit] = hit_sphere,
-	[Parse] = parse_sphere,
 	[Debug] = debug_sphere};
 
 	return (functions[func]);
