+++ src/file/file.h
@@ -6,19 +6,19 @@
 /*   By: yonshin <yonshin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/31 04:18:35 by minjungk          #+#    #+#             */
-/*   Updated: 2023/06/09 23:56:45 by minjungk         ###   ########.fr       */
+/*   Updated: 2023/07/03 22:47:55 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FILE_H
 # define FILE_H
 # include <fcntl.h>
-# include "Element/AmbientLight/ambient_light.h"
-# include "Element/Camera/camera.h"
-# include "Element/Light/light.h"
-# include "Element/Plane/plane.h"
-# include "Element/Sphere/sphere.h"
-# include "Element/Cylinder/cylinder.h"
+# include "../Element/AmbientLight/ambient_light.h"
+# include "../Element/Camera/camera.h"
+# include "../Element/Light/light.h"
+# include "../Element/Plane/plane.h"
+# include "../Element/Sphere/sphere.h"
+# include "../Element/Cylinder/cylinder.h"
 
 struct s_parse_info
 {
