+++ src/file/parse_ambient_light.c
@@ -1,34 +1,32 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_light.c                                      :+:      :+:    :+:   */
+/*   parse_ambient_light.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: minjungk <minjungk@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/31 05:37:19 by minjungk          #+#    #+#             */
-/*   Updated: 2023/07/03 22:41:05 by minjungk         ###   ########.fr       */
+/*   Created: 2023/05/31 05:35:38 by minjungk          #+#    #+#             */
+/*   Updated: 2023/07/03 22:48:35 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "light.h"
+# include "../Element/AmbientLight/ambient_light.h"
 
-int	parse_light(void *param, char **argv)
+int	parse_ambient_light(void *param, char **argv)
 {
-	char					*remain;
-	struct s_light *const	content = param;
+	char							*remain;
+	struct s_ambient_light *const	content = param;
 
 	if (!argv || !argv[0] || !argv[1])
 		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
-	content->type = Light;
-	if (parse_vector3(&content->raw.coordinate, argv[0], AllScope))
-		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	content->raw.ratio = ft_strtof(argv[1], &remain);
+	content->type = AmbientLight;
+	content->raw.ratio = ft_strtof(argv[0], &remain);
 	if (errno || (*remain != '\0' && ft_strchr("\r\n", *remain) == NULL))
 		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
+	if (parse_rgb(&content->raw.rgb, argv[1]))
+		return (ft_error(__func__, __FILE__, __LINE__, 0));
 	if (!(0.0 <= content->raw.ratio && content->raw.ratio <= 1.0))
 		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
-	if (argv[2] && parse_rgb(&content->raw.rgb, argv[2]))
-		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	debug_light(content);
+	debug_ambient_light(content);
 	return (EXIT_SUCCESS);
 }
