+++ src/file/parse_sphere.c
@@ -1,34 +1,32 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_camera.c                                     :+:      :+:    :+:   */
+/*   parse_sphere.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: minjungk <minjungk@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/31 05:36:34 by minjungk          #+#    #+#             */
-/*   Updated: 2023/07/03 22:40:51 by minjungk         ###   ########.fr       */
+/*   Created: 2023/05/31 05:37:44 by minjungk          #+#    #+#             */
+/*   Updated: 2023/07/03 22:49:01 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "camera.h"
+# include "../Element/Sphere/sphere.h"
 
-int	parse_camera(void *param, char **argv)
+int	parse_sphere(void *param, char **argv)
 {
 	char					*remain;
-	struct s_camera *const	content = param;
+	struct s_sphere *const	content = param;
 
 	if (!argv || !argv[0] || !argv[1] || !argv[2])
 		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
-	content->type = Camera;
+	content->type = Sphere;
 	if (parse_vector3(&content->raw.coordinate, argv[0], AllScope))
 		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	if (parse_vector3(&content->raw.axis, argv[1], UnitScope))
-		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	content->raw.fov = ft_strtof(argv[2], &remain);
+	content->raw.diameter = ft_strtof(argv[1], &remain);
 	if (errno || (*remain != '\0' && ft_strchr("\r\n", *remain) == NULL))
 		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
-	if (!(0.0 <= content->raw.fov && content->raw.fov <= 180.0))
-		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
-	debug_camera(content);
+	if (parse_rgb(&content->raw.rgb, argv[2]))
+		return (ft_error(__func__, __FILE__, __LINE__, 0));
+	debug_sphere(content);
 	return (EXIT_SUCCESS);
 }
